pyproject.toml
[tool.poetry]
name = "courswork4"
version = "0.1.0"
description = ""
authors = ["ryfe <denis.koptelev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
utils = "^1.0.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.isort]
line_length = 120